this app source editor tsd is committed into the repo, so we don't need it before a build
