+++ frontend/build.mjs
@@ -9,13 +9,11 @@ import {
     buildInParallel,
     copyIndexHtml,
 } from './utils.mjs'
-import fse from 'fs-extra'
 
 export const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
 startDevServer(__dirname)
 copyPublicFolder(path.resolve(__dirname, 'public'), path.resolve(__dirname, 'dist'))
-writeSourceCodeEditorTypes()
 writeIndexHtml()
 writeSharedDashboardHtml()
 
@@ -86,32 +84,6 @@ await buildInParallel(
         },
     }
 )
-
-export function writeSourceCodeEditorTypes() {
-    const readFile = (p) => {
-        try {
-            return fse.readFileSync(path.resolve(__dirname, p), { encoding: 'utf-8' })
-        } catch (e) {
-            if (isDev) {
-                console.warn(
-                    `🙈 Didn't find "${p}" for the app source editor. Build it with: yarn build:packages:types`
-                )
-            } else {
-                throw e
-            }
-        }
-    }
-    const types = {
-        '@types/react/index.d.ts': readFile('../node_modules/@types/react/index.d.ts'),
-        '@types/react/global.d.ts': readFile('../node_modules/@types/react/global.d.ts'),
-        '@types/kea/index.d.ts': readFile('../node_modules/kea/lib/index.d.ts'),
-        '@posthog/apps-common/index.d.ts': readFile('./@posthog/apps-common/dist/index.d.ts'),
-        '@posthog/lemon-ui/index.d.ts': readFile('./@posthog/lemon-ui/dist/index.d.ts'),
-    }
-    const packagesJsonFile = path.resolve(__dirname, './src/scenes/plugins/source/types/packages.json')
-    fse.mkdirpSync(path.dirname(packagesJsonFile))
-    fse.writeFileSync(packagesJsonFile, JSON.stringify(types, null, 4) + '\n')
-}
 
 export function writeIndexHtml(chunks = {}, entrypoints = []) {
     copyIndexHtml(__dirname, 'src/index.html', 'dist/index.html', 'index', chunks, entrypoints)
