+++ package.json
@@ -26,11 +26,12 @@
         "start-ch-dev": "concurrently -n DOCKER,ESBUILD,TYPEGEN -c red,blue,green \"docker-compose -f docker-compose.dev.yml pull && CH_WEB_SCRIPT=./ee/bin/docker-ch-dev-backend docker-compose -f docker-compose.dev.yml up\" \"yarn run start-http --host 0.0.0.0\" \"yarn run typegen:watch\"",
         "clear-ch-dev": "docker compose -f docker-compose.dev.yml stop && docker compose -f docker-compose.dev.yml rm -v && docker compose -f docker-compose.dev.yml down",
         "clean": "rm -rf frontend/dist && mkdir frontend/dist",
-        "build": "yarn copy-scripts && yarn build:packages && yarn build:esbuild",
+        "build": "yarn copy-scripts && yarn build:esbuild",
         "build:esbuild": "node frontend/build.mjs",
-        "build:packages": "yarn build:packages:apps-common && yarn build:packages:lemon-ui",
-        "build:packages:apps-common": "cd frontend/@posthog/apps-common && yarn && yarn build",
-        "build:packages:lemon-ui": "cd frontend/@posthog/lemon-ui && yarn && yarn build",
+        "editor:update-tsd": "node frontend/editor-update-tsd.mjs",
+        "packages:build": "yarn packages:build:apps-common && yarn packages:build:lemon-ui",
+        "packages:build:apps-common": "cd frontend/@posthog/apps-common && yarn && yarn build",
+        "packages:build:lemon-ui": "cd frontend/@posthog/lemon-ui && yarn && yarn build",
         "prettier": "prettier --write \"./**/*.{js,mjs,ts,tsx,json,yaml,yml,css,scss}\"",
         "prettier:check": "prettier --check \"./**/*.{js,mjs,ts,tsx,json,yaml,yml,css,scss}\"",
         "typescript:check": "tsc --noEmit && echo \"No errors reported by tsc.\"",
